requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==4.2.14
django-admin-interface==0.28.8
django-admin-persian-fonts==0.2
django-colorfield==0.11.0
django-import-export==4.1.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.4.0
pyTelegramBotAPI==4.21.0
python-slugify==8.0.4
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
text-unidecode==1.3
urllib3==2.2.2
whitenoise==6.7.0
yarl==1.9.4
